requirements.txt
streamlit
google-generativeai
python-dotenv
pathlib
pyttsx3
gtts
pydub
